refactor(git): remove remote-write retry logic
This removes complexity required for dedicated retries. Instead we can
fallback on the periodic sync retry that we have going anyway.
