+++ src/main/resources/application-default.yml
@@ -14,8 +14,3 @@ scheduling:
   push:
     initial-delay: 10000
     pause-delay: 5000
-
-retry:
-  remote-write:
-    attempts: 3
-    backoff-delay: 2000
